+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 8.4
 - Add File Attached to Incomes in Incomes.php & Accounting/functions.inc.php
 - Add File Attached to Expenses in Expenses.php
 - SQL ACCOUNTING_INCOMES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
+- Fix SQL error when no user in session in Template.fnc.php
 
 Changes in 8.3.1
 ----------------